+++ src/qibocal/protocols/characterization/__init__.py
@@ -2,8 +2,8 @@ from enum import Enum
 
 from .classification import single_shot_classification
 from .qubit_spectroscopy import qubit_spectroscopy
-from .rabi import rabi_amplitude
-from .rabi_length import rabi_length
+from .rabi.amplitude import rabi_amplitude
+from .rabi.length import rabi_length
 from .ramsey import ramsey
 from .resonator_punchout import resonator_punchout
 from .resonator_spectroscopy import resonator_spectroscopy
